pyproject.toml
[project]
name = "pms-telegram-bot"
version = "0.1.0"
description = "Telegram bot that submits iQube PMS daily logs on behalf of college users by replaying their Microsoft OAuth login in a per-user Playwright browser."
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[webhooks]>=21.6",
    "playwright>=1.48",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "cryptography>=43",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
